Build LLVM toolchain based on OSP offer
This patch allows to build LLVM based on the experimental OS porting
(OSP) toolchain.  It builds the binutils, clang/llvm, and newlib from
github sources.  Note that compiler-rt is missing here in order to
have a complete OSP toolchain working but the toolchain is able to
generate asm from C.